kye pls

+++ Commando.js
@@ -60,7 +60,7 @@ client.on('error', winston.error)
 		if (message.channel.type === 'dm') return;
 
 		if (message.guild.id === '222078108977594368' && !message.member.roles.has('242700009961816065') && /https:\/\/discord\.gg\/\w+/.test(message.content)) {
-			message.delete();
+			if (message.deletable && !message.author.id !== client.user.id) message.delete();
 			message.reply('Please do not post invite links on this server. If you wish to give invite links, do so in direct messages.');
 		}
 		const channelLocks = client.provider.get(message.guild.id, 'locks', []);
